+++ src/App.tsx
@@ -13,6 +13,8 @@ import { ChainbridgeProvider } from "./Contexts/ChainbridgeContext";
 import AppWrapper from "./Layouts/AppWrapper";
 import { chainbridgeConfig } from "./chainbridgeConfig";
 import { NetworkProvider } from "./Contexts/NetworkContext";
+import { Web3Provider } from "@chainsafe/web3-context";
+import { utils } from "ethers";
 
 if (
   process.env.NODE_ENV === "production" &&
@@ -26,6 +28,13 @@ if (
 }
 
 const App: React.FC<{}> = () => {
+  const tokens = chainbridgeConfig.chains.reduce((tca, bc) => {
+    return {
+      ...tca,
+      [bc.networkId]: bc.tokens,
+    };
+  }, {});
+
   return (
     <ErrorBoundary
       fallback={({ error, componentStack, eventId, resetError }) => (
@@ -49,7 +58,23 @@ const App: React.FC<{}> = () => {
       <ThemeSwitcher themes={{ light: lightTheme }}>
         <CssBaseline />
         <ToasterProvider autoDismiss>
-          <NetworkProvider>
+          <Web3Provider
+            tokensToWatch={tokens}
+            onboardConfig={{
+              dappId: process.env.REACT_APP_BLOCKNATIVE_DAPP_ID,
+              walletSelect: {
+                wallets: [{ walletName: "metamask", preferred: true }],
+              },
+              subscriptions: {
+                network: (network) => console.log("chainId: ", network),
+                balance: (amount) =>
+                  console.log("balance: ", utils.formatEther(amount)),
+              },
+            }}
+            checkNetwork={false}
+            gasPricePollingInterval={120}
+            gasPriceSetting="fast"
+          >
             <ChainbridgeProvider>
               <Router>
                 <AppWrapper>
@@ -57,7 +82,7 @@ const App: React.FC<{}> = () => {
                 </AppWrapper>
               </Router>
             </ChainbridgeProvider>
-          </NetworkProvider>
+          </Web3Provider>
         </ToasterProvider>
       </ThemeSwitcher>
     </ErrorBoundary>
